// src/schemas/auth/sign-up-schema.ts
import z from "zod";

export const SignUpSchema = z
  .object({
    firstName: z.string().min(3, "First name must be at least 3 characters"),
    lastName: z.string().min(3, "Last name must be at least 3 characters"),
    username: z.string().min(3, "Username must be at least 3 characters"),
    email: z.email("Email format is invalid"),
    password: z.string().min(6, "Password must be at least 6 characters"),
    confirmPassword: z.string(),
  })
  .refine((data) => data.password === data.confirmPassword, {
    message: "Passwords don't match",
    path: ["confirmPassword"],
  });